Chatbot-Service/requirements.txt
flask
chromadb
requests
python-dotenv
pydantic
sentence-transformers